RMINC Masked Value

RMINC_MASKED_VALUE is an object with the S3 class "RMINC_MASKED_VALUE" with NA value used to differentiate whether an NA was generated by the user or by an RMINC masking procedure. The value, which will be written out with mincWriteVolume, can be changed to zero for example with setRMINCMaskedValue.