Read all CIVET files into R

Parses outputs from CIVET pipeline

Usage

civet.readAllCivetFiles(atlasFile, gf, civetVersion = "1.1.12")

Arguments

atlasFile

Character path to a key to the atlas used when running civet. the key should be a comma separated file with a header and the following form
Column 1: Numeric label
Column 3: Corresponding structure

gf

Data Frame containing list of all CIVET file names, and where results will be stored requires gf to have an element (column) called CIVETFILES which is a data.frame containing paths to the civetFiles, typically generated with civet.getAllFilenames

civetVersion

The version of CIVET that produced the files.

Value

Returns gf augmented with additional columns containing

lobeArea40mm: A subjects by region matrix of lobe areas parcellated by atlas region

lobeThickness: As above, but for thicknesses

lobeVolume: As above, but for volumes

GI: A subjects by 6 matrix. 6 columns are left hemisphere gyrification indices for the gray matter surface, white matter surface, midsurface of the two, followed by the same indices for the two

BrainVolume: A subjects by 3 matrix. Three columns are CSF volume, grey matter and white matter respectively

midSurfaceNativeArea: A subjects by vertices matrix of mid-surface areas

SurfaceNativeVolume: As above, but for native space volumes

nativeRMS_RSLtlink20mm: As above, but for RMS_RSL tlink 20mm thicknesses

nativeRMStlink20mm: As above, but for RMS tlink 20mm thicknesses

Details

Prior to running, civet.getAllFilenames may be called to generate the input argument gf . This function will extract the following information from the CIVET pipeline: Lobe Area (40 mm), Lobe Thickness, Lobe Volume, GI, Mid Surface Native Area, Surface Native Volume, Brain Volume Native RMS RSL tLink (20mm), Native RMS tLink (20 mm)

See also

Examples

if (FALSE) { # \dontrun{
getRMINCTestData()
gf = read.csv("/tmp/rminctestdata/CIVET_TEST.csv")
gf = civet.getAllFilenames(gf,"ID","TEST","/tmp/rminctestdata/CIVET", TRUE, "1.1.12")
gf = civet.readAllCivetFiles("/tmp/rminctestdata/AAL.csv",gf)
} # }