Find connected components of a object graph

Given vertex data and a threshold, find the connected component of a graph representing a bic_obj

Usage

connected_components(obj_graph, data_map, data_range)

Arguments

obj_graph

The igraph object generated by obj_to_graph

data_map

A set of data corresponding to the vertices

data_range

A two-component vector giving lower and upper bounds for vertices to include when generating connected components

Value

a data frame of vertices after thresholding indicating the vertex id, the cluster it's a member of, and the size of its cluster