Plot Slice Along Each Axis

Show a slice from each axis of minc volume

Usage

mincTriplanarSlicePlot(
  anatomy,
  statistics,
  slice = NULL,
  layoutMatrix = NULL,
  ...
)

Arguments

anatomy

a mincArray object containing the source anatomy

statistics

a mincArray object containing a statistic to overlay

slice

3-component vector indicating which slice along each axis

layoutMatrix

A matrix describing the layout for the plots typically produced by layout

...

extra parameters to be passed to mincPlotAnatAndStatsSlice

Value

invisible NULL