Two tailed version of pt

Generate two tailed probabilities from the cummulative t distribution function with a given number of degrees of freedom.

Usage

pt2(q, df, log.p = FALSE)

Arguments

q

a vector of quantiles

df

degrees of freedom

log.p

whether to return the natural logarithm of p

Value

A vector of probabilities

See also